Simplify and make use of forked module
This module was originally created to extend Ben Manu's Silverstripe
LeafletField in order to add Google Mutant functionailty on the
front-end. This is done predominantly through alteration of the included
scripts and styles in the `Requirements` static function calls.

As the function that makes these calls is protected, we can simply
achieve this by subclassing rather than recreating a whole module. To
this end the module is marked as abandoned but left in a working state
for backwards compatibility, allowing time to change project level class
references.